// tools/api-reference/generated-file.ts
export interface GeneratedApi {
  readonly references: Readonly<Record<string, ApiReferenceData>>;
  readonly referencesByRoute: Readonly<Record<string, readonly string[]>>;
}